Refactor code again after following along with lecture, see below for errors
Remember that constructor functions within a class are called immediately
when a new Object is being initialized from the Class, so if you want a
function to execute immediately upon page load, you can place that
function call inside of the constructor, alongside the class's other
properties;

Note however, that specifically for EVENT LISTENER FUNCTIONS, the
'this' keyword within a event listener function points to the DOM
element on which it was called, and so is undefined by definition.
In order to keep the 'this' keyword pointing to the Class Object in
which the operations are taking place, you will need to 'BIND' the
'this' keyword to the Object by using ' handler.bind(this) ' , or
' this.handler.bind(this) ' if defining in the constructor function
within a Class function.